extracted all other input fields
